benchmarks/full-reduced/video-analytics/video-full/requirements.txt
parliament-functions==0.1.0
environs==9.3.2
pycapnp==1.3.0
opencv-python-headless
requests
tracing
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
torchvision==0.12.0+cpu
torchaudio
